+++ CI/SCRIPTS/test_vectors.py
@@ -25,7 +25,7 @@ from fiona.errors import DriverError
 from rasterio import CRS
 from shapely import wkt
 
-from CI.SCRIPTS.script_utils import s3_env, vectors_path
+from CI.SCRIPTS.script_utils import files_path, s3_env, vectors_path
 from sertit import ci, files, vectors
 from sertit.vectors import WGS84
 
@@ -179,6 +179,7 @@ def test_copy():
 
 @s3_env
 def test_utm_context_manager():
+    """Test UTM context manager."""
     warning_msg = r"Geometry is in a geographic CRS. Results from 'centroid' are likely incorrect.*$"
 
     # Open a geographic vector
@@ -212,3 +213,27 @@ def test_utm_context_manager():
 
     # Assert the column still exists
     assert vect["centroid_utm"].equals(c2)
+
+
+def test_read_archived():
+    """Test archived vectors"""
+    landsat = "LM05_L1TP_200030_20121230_20200820_02_T2_CI"
+    map_overlay = "map-overlay.kml"
+    map_overlay_regex = ".*{0}".format(map_overlay.replace(".", r"\."))
+    map_overlay_extracted_path = files_path() / landsat / map_overlay
+    zip_landsat = files_path() / f"{landsat}.zip"
+    tar_landsat = files_path() / f"{landsat}.tar"
+
+    map_overlay_extracted = vectors.read(map_overlay_extracted_path)
+
+    ci.assert_geom_equal(
+        map_overlay_extracted, vectors.read(f"{zip_landsat}!{landsat}/{map_overlay}")
+    )
+    ci.assert_geom_equal(
+        map_overlay_extracted,
+        vectors.read(zip_landsat, archive_regex=map_overlay_regex),
+    )
+    ci.assert_geom_equal(
+        map_overlay_extracted,
+        vectors.read(tar_landsat, archive_regex=map_overlay_regex),
+    )